requirements.txt
Flask==3.0.3
requests==2.32.3
python-telegram-bot==20.7
python-dotenv==1.0.0
werkzeug==2.3.7
aiohttp==3.8.5